latest commit

+++ src/main/java/com/learing/myproject/todo/service/TaskServiceImpl.java
@@ -7,9 +7,16 @@ import com.learing.myproject.todo.repository.TaskListRepository;
 import com.learing.myproject.todo.repository.TaskRepository;
 import com.learing.myproject.todo.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.stereotype.Service;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
+@Service
 public class TaskServiceImpl implements TaskService{
 
     @Autowired
@@ -21,6 +28,9 @@ public class TaskServiceImpl implements TaskService{
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private MongoTemplate mongoTemplate;
+
     @Override
     public List<Task> findAllTasks(String username) {
         List<Task> allTasks = new ArrayList<>();
@@ -30,21 +40,15 @@ public class TaskServiceImpl implements TaskService{
         for(TaskList taskList : allTaskLists) {
             allTaskIds.addAll(taskList.getTaskIds());
         }
-        for(String taskId : allTaskIds) {
-            Task theTask = findById(taskId);
-            allTasks.add(theTask);
-        }
+        allTasks = taskRepository.findAllById(allTaskIds);
         allTasks.sort(Comparator.comparing(Task::getDueDate).reversed());
         return allTasks;
     }
 
     @Override
     public Map<String, List<String>> getTasksByLists(String username) {
-
         User userEntity = findUserByUsername(username);
-
         List<String> taskListIds = userEntity.getTaskListIds();
-
         Map<String, List<String>> tasksByLists = new HashMap<>();
 
         for(String ListId: taskListIds){
@@ -55,7 +59,6 @@ public class TaskServiceImpl implements TaskService{
                     .toList();
             tasksByLists.put(taskList.getListName(), taskNames);
         }
-
         return tasksByLists;
     }
 
@@ -72,9 +75,6 @@ public class TaskServiceImpl implements TaskService{
         return tasksByListName;
     }
 
-
-
-
     @Override
     public Task createTask(String username, Task task, String listName) {
         User user = findUserByUsername(username);
@@ -90,28 +90,64 @@ public class TaskServiceImpl implements TaskService{
     public TaskList createTaskList(String username, TaskList taskList) {
         User user = findUserByUsername(username);
         taskList.setUserId(user.getId());
+        taskList.setTaskIds(new ArrayList<>());
+        taskList = taskListRepository.save(taskList);
         user.getTaskListIds().add(taskList.getId());
         user = userRepository.save(user);
-        TaskList newTaskList = taskListRepository.save(taskList);
-        return newTaskList;
+        return taskList;
     }
 
     @Override
-    public Task updateTask(String id, Task task) {
+    public Task updateTask(String username, String id, Task task) {
+        User user = findUserByUsername(username);
+        if(!userHasAccess(user, id))
+            throw new AccessDeniedException("Access Denied");
         Task theTask = findById(id);
-        
         theTask.setTitle(task.getTitle());
         theTask.setDescription(task.getDescription());
         return taskRepository.save(task);
     }
 
     @Override
-    public void deleteTask(String id) {
+    public void deleteTask(String username, String id) {
+        User user = findUserByUsername(username);
+        if(!userHasAccess(user, id))
+            throw new AccessDeniedException("Access Denied");
         Task task = findById(id);
         taskRepository.delete(task);
     }
 
     @Override
+    public Task getTaskById(String username, String id) {
+        User user = findUserByUsername(username);
+        if(!userHasAccess(user, id))
+            throw new AccessDeniedException("Access Denied");
+        Task task = findById(id);
+        return task;
+    }
+
+    @Override
+    public List<Task> sortAndFilter(String username, String sortField, String filterField) {
+       List<Task> allTasks = findAllTasks(username);
+       return groupAndSort(allTasks, sortField, filterField);
+    }
+
+    @Override
+    public List<Task> searchTasks(String username, String searchText) {
+        User user = findUserByUsername(username);
+        List<String> taskListIds = user.getTaskListIds();
+        Criteria criteria = new Criteria().andOperator(
+                Criteria.where("taskListId").in(taskListIds),
+                new Criteria().orOperator(
+                        Criteria.where("title").regex(searchText),
+                        Criteria.where("description").regex(searchText)
+                )
+        );
+        Query query = new Query(criteria);
+        List<Task> tasks = mongoTemplate.find(query, Task.class);
+        return tasks;
+    }
+
     public Task findById(String id) {
         return taskRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Task not found with id: " + id));
@@ -122,5 +158,54 @@ public class TaskServiceImpl implements TaskService{
                 .orElseThrow(() -> new RuntimeException("User not found with username " + username));
     }
 
-    public boolean userHasAccess ()
+    public boolean userHasAccess (User user, String id){
+        HashSet<String> taskIdSet = new HashSet<>();
+        List<TaskList> allTaskLists = taskListRepository.findByUserId(user.getId());
+        for(TaskList taskList : allTaskLists) {
+            taskIdSet.addAll(taskList.getTaskIds());
+        }
+        return taskIdSet.contains(id);
+    }
+
+    public List<Task> groupAndSort(List<Task> tasks, String sortField, String filterField) {
+        Map<Object, List<Task>> groupedTasks = new HashMap<>();
+        switch (filterField) {
+            case "dueDate":
+                groupedTasks = tasks.stream().collect(Collectors.groupingBy(Task::getDueDate));
+                break;
+            case "category":
+                groupedTasks = tasks.stream().collect(Collectors.groupingBy(Task::getCategory));
+                break;
+            case "completed":
+                groupedTasks = tasks.stream().collect(Collectors.groupingBy(Task::isCompleted));
+                break;
+            case "priority":
+                groupedTasks = tasks.stream().collect(Collectors.groupingBy(Task::getPriority));
+                break;
+            default:
+                System.out.println("Invalid groupBy value. Supported values are: dueDate, category, completed, priority");
+                return tasks;
+        }
+
+        // Sorting logic based on the value of sortBy
+        Comparator<Task> comparator;
+        switch (sortField) {
+            case "dueDate" -> comparator = Comparator.comparing(Task::getDueDate);
+            case "priority" -> comparator = Comparator.comparing(Task::getPriority);
+            default -> {
+                System.out.println("Invalid sortBy value. Supported values are: dueDate, priority");
+                return tasks;
+            }
+        }
+
+        // Sorting each group
+        groupedTasks.forEach((key, value) -> value.sort(comparator));
+
+        // Flattening the map to retrieve sorted tasks
+        List<Task> sortedAndGroupedTasks = groupedTasks.values().stream()
+                .flatMap(Collection::stream)
+                .collect(Collectors.toList());
+
+        return sortedAndGroupedTasks;
+    }
 }
